+++ src/aiosmtplib/__init__.py
@@ -14,6 +14,7 @@ from .api import send
 from .errors import (
     SMTPAuthenticationError,
     SMTPConnectError,
+    SMTPConnectResponseError,
     SMTPConnectTimeoutError,
     SMTPDataError,
     SMTPException,
@@ -26,15 +27,13 @@ from .errors import (
     SMTPSenderRefused,
     SMTPServerDisconnected,
     SMTPTimeoutError,
-    SMTPConnectResponseError,
 )
 from .response import SMTPResponse
 from .smtp import SMTP
 from .typing import SMTPStatus, SMTPTokenGenerator
 
-
 __title__ = "aiosmtplib"
-__version__ = "5.1.0"
+__version__ = "5.1.1"
 __author__ = "Cole Maclean"
 __license__ = "MIT"
 __copyright__ = "Copyright 2022 Cole Maclean"